# frontend/openapi.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Labour Tracker",
        "version": "0.1.0"
    },
    "paths": {
...
        "/api/v1/subscription-management/remove-subscriber": {
            "put": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "Remove Subscriber",
                "operationId": "remove_subscriber_api_v1_subscription_management_remove_subscriber_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RemoveSubscriberRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/subscription-management/block-subscriber": {
            "put": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "Block Subscriber",
                "operationId": "block_subscriber_api_v1_subscription_management_block_subscriber_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/BlockSubscriberRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/subscription-management/update-role": {
            "put": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "Update Role",
                "operationId": "update_role_api_v1_subscription_management_update_role_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateRoleRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
        "/api/v1/subscription-management/update-contact-methods": {
            "put": {
                "tags": [
                    "Subscription Management"
                ],
                "summary": "Update Contact Methods",
                "operationId": "update_contact_methods_api_v1_subscription_management_update_contact_methods_put",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/UpdateContactMethodsRequest"
                            }
                        }
                    },
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/SubscriptionResponse"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ExceptionSchema"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "HTTPBearer": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "BlockSubscriberRequest": {
                "properties": {
                    "subscription_id": {
                        "type": "string",
                        "title": "Subscription Id"
                    }
                },
                "type": "object",
                "required": [
                    "subscription_id"
                ],
                "title": "BlockSubscriberRequest"
            },
...
            "ExceptionSchema": {
                "properties": {
                    "description": {
                        "type": "string",
                        "title": "Description"
                    }
                },
                "type": "object",
                "required": [
                    "description"
                ],
                "title": "ExceptionSchema"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
...
            "RemoveSubscriberRequest": {
                "properties": {
                    "subscription_id": {
                        "type": "string",
                        "title": "Subscription Id"
                    }
                },
                "type": "object",
                "required": [
                    "subscription_id"
                ],
                "title": "RemoveSubscriberRequest"
            },
...
            "SubscriptionDTO": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "labour_id": {
                        "type": "string",
                        "title": "Labour Id"
                    },
                    "birthing_person_id": {
                        "type": "string",
                        "title": "Birthing Person Id"
                    },
                    "subscriber_id": {
                        "type": "string",
                        "title": "Subscriber Id"
                    },
                    "role": {
                        "type": "string",
                        "title": "Role"
                    },
                    "status": {
                        "type": "string",
                        "title": "Status"
                    },
                    "contact_methods": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Contact Methods"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "labour_id",
                    "birthing_person_id",
                    "subscriber_id",
                    "role",
                    "status",
                    "contact_methods"
                ],
                "title": "SubscriptionDTO"
            },
            "SubscriptionResponse": {
                "properties": {
                    "subscription": {
                        "$ref": "#/components/schemas/SubscriptionDTO"
                    }
                },
                "type": "object",
                "required": [
                    "subscription"
                ],
                "title": "SubscriptionResponse"
            },
...
            "UpdateContactMethodsRequest": {
                "properties": {
                    "subscription_id": {
                        "type": "string",
                        "title": "Subscription Id"
                    },
                    "contact_methods": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Contact Methods"
                    }
                },
                "type": "object",
                "required": [
                    "subscription_id",
                    "contact_methods"
                ],
                "title": "UpdateContactMethodsRequest"
            },
            "UpdateRoleRequest": {
                "properties": {
                    "subscription_id": {
                        "type": "string",
                        "title": "Subscription Id"
                    },
                    "role": {
                        "type": "string",
                        "title": "Role"
                    }
                },
                "type": "object",
                "required": [
                    "subscription_id",
                    "role"
                ],
                "title": "UpdateRoleRequest"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            }
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}